Add allowDiskUse to aggregate call
In the get_first_snapshot_times() function in the stakeholder_info blueprint
there is an aggregate. Today (2020-12-22) it was reported that users were
seeing an Internal Server Error when attempting to use this functionality. Upon
investigating I found the following error in the log:

OperationFailure: Sort exceeded memory limit of 104857600 bytes, but did not
opt in to external sorting.

We surmised that adding the allowDiskUse=True kwarg would resolve this. After
adding it and restarting the service, further requests had no issue.