Fix up to have at least one pytest test
When running on GitHub Actions, pytest has an exit code of 5 because there are
no tests. We need an exit code of 0 to pass, so I have added some barebones
tests for ncats_webd.launch as it was the easiest to target.
I also re-added *.pyc to the .gitignore because of our Python 2 restriction.