Update numpy and pandas versions to match those in cisagov/cyhy-reports
Since this repo is a dependency of cisagov/cyhy-reports, and the
versions of numpy and pandas are pinned in that repository, the
versions must also be pinned here.  If not, then cyhy-reports installs
one version of these Python packages, ncats-webd installs another, and
cisagov/ansible-role-cyhy-reports is not idempotent.  More
importantly, the version of these Python packages that actually gets
installed via cisagov/ansible-role-cyhy-reports is undetermined.